Module awc::error

HTTP client errors

Structs

HeaderValue

Represents an HTTP header field value.

HttpError

A generic “error” for HTTP connections

StatusCode

An HTTP status code (status-code in RFC 7230 et al.).

Enums

ConnectError

A set of errors that can occur while connecting to an HTTP host

FreezeRequestError

A set of errors that can occur during freezing a request

InvalidUrl

JsonPayloadError

A set of errors that can occur during parsing json payloads

PayloadError

A set of errors that can occur during payload parsing.

SendRequestError

A set of errors that can occur during request sending and response reading

WsClientError

Websocket client error

WsHandshakeError

WebSocket handshake errors

WsProtocolError

WebSocket protocol errors.