Struct awc::error::HttpError

pub struct HttpError { /* private fields */ }

A generic โerrorโ for HTTP connections

This error type is less specific than the error returned from other functions in this crate, but all other errors can be converted to this error. Consumers of this crate can typically consume and work with this form of error for conversions with the ? operator.

Implementations

impl Error

pub fn is<T>(&self) -> bool
where T: Error + 'static,

Return true if the underlying error has the same type as T.

pub fn get_ref(&self) -> &(dyn Error + 'static)

Return a reference to the lower level, inner error.