Struct awc::ClientBuilder

pub struct ClientBuilder { /* private fields */ }

An HTTP Client builder

This type can be used to construct an instance of Client through a builder-like pattern.

Implementations

impl ClientBuilder

pub fn new() -> Self

pub fn connector<T>(self, connector: T) -> Self where
T: Service<Request = HttpConnect, Error = ConnectError> + 'static,
T::Response: Connection,
<T::Response as Connection>::Future: 'static,
T::Future: 'static,

Use custom connector service.

pub fn timeout(self, timeout: Duration) -> Self

Set request timeout

Request timeout is the total time before a response must be received. Default value is 5 seconds.

pub fn disable_timeout(self) -> Self

Disable request timeout.

pub fn disable_redirects(self) -> Self

Do not follow redirects.

Redirects are allowed by default.

pub fn max_http_version(self, val: Version) -> Self

Maximum supported http major version Supported versions http/1.1, http/2

pub fn initial_window_size(self, size: u32) -> Self

Indicates the initial window size (in octets) for HTTP2 stream-level flow control for received data.

The default value is 65,535 and is good for APIs, but not for big objects.

pub fn initial_connection_window_size(self, size: u32) -> Self

Indicates the initial window size (in octets) for HTTP2 connection-level flow control for received data.

The default value is 65,535 and is good for APIs, but not for big objects.

pub fn max_redirects(self, num: usize) -> Self

Set max number of redirects.

Max redirects is set to 10 by default.

pub fn no_default_headers(self) -> Self

Do not add default request headers. By default Date and User-Agent headers are set.

pub fn header<K, V>(self, key: K, value: V) -> Self where
HeaderName: TryFrom<K>,
<HeaderName as TryFrom<K>>::Error: Debug + Into<HttpError>,
V: IntoHeaderValue,
V::Error: Debug,

Add default header. Headers added by this method get added to every request.

pub fn basic_auth<U>(self, username: U, password: Option<&str>) -> Self where
U: Display,

Set client wide HTTP basic authorization header

pub fn bearer_auth<T>(self, token: T) -> Self where
T: Display,

Set client wide HTTP bearer authentication header

pub fn finish(self) -> Client

Finish build process and create Client instance.