Crate awc

awc is a HTTP and WebSocket client library built on the Actix ecosystem.

Making a GET request

let mut client = awc::Client::default();
let response = client.get("http://www.rust-lang.org") // <- Create request builder
    .insert_header(("User-Agent", "Actix-web"))
    .send()                                            // <- Send http request
    .await?;

 println!("Response: {:?}", response);

Making POST requests

Raw body contents

let mut client = awc::Client::default();
let response = client.post("http://httpbin.org/post")
    .send_body("Raw body contents")
    .await?;

Forms

let params = [("foo", "bar"), ("baz", "quux")];

let mut client = awc::Client::default();
let response = client.post("http://httpbin.org/post")
    .send_form(&params)
    .await?;

JSON

let request = serde_json::json!({
    "lang": "rust",
    "body": "json"
});

let mut client = awc::Client::default();
let response = client.post("http://httpbin.org/post")
    .send_json(&request)
    .await?;

Response Compression

All official and common content encoding codecs are supported, optionally.

The Accept-Encoding header will automatically be populated with enabled codecs and added to outgoing requests, allowing servers to select their Content-Encoding accordingly.

Feature flags enable these codecs according to the table below. By default, all compress-* features are enabled.

Feature	Codecs

compress-brotli	brotli

compress-gzip	gzip, deflate

compress-zstd	zstd

WebSocket support

use futures_util::{sink::SinkExt, stream::StreamExt};
let (_resp, mut connection) = awc::Client::new()
    .ws("ws://echo.websocket.org")
    .connect()
    .await?;

connection
    .send(awc::ws::Message::Text("Echo".into()))
    .await?;
let response = connection.next().await.unwrap()?;

Re-exports

pub use cookie;

Modules

error

HTTP client errors

http

Various HTTP related types.

middleware

test

Test helpers for actix http client to use during testing.

ws

Websockets client

Structs

Client

An asynchronous HTTP and WebSocket client.

ClientBuilder

An HTTP Client builder

ClientRequest

An HTTP Client request builder

ClientResponse

Client Response

Connector

Manages HTTP client network connectivity.

FrozenClientRequest

FrozenClientRequest struct represents cloneable client request. It could be used to send same request multiple times.

FrozenSendBuilder

Builder that allows to modify extra headers.

JsonBody

Response’s payload json parser, it resolves to a deserialized T value.

MessageBody

Future that resolves to a complete HTTP message body.

Enums

ConnectRequest

ConnectResponse

SendClientRequest

Future that sends request’s payload and resolves to a server response.

Type Definitions

BoxConnectorService

BoxedSocket